Rename constants module to settings
